shryltyan
Shryltyan is a fictional character from the animated television series "Avatar: The Last Airbender." She is portrayed by actress Zoe Saldaña. The character is a skilled waterbender and the daughter of the Avatar, Aang. Shryltyan is known for her strong sense of justice and her dedication to protecting her people, the Water Tribes.